Re-structure of REST API packages and shifted Boot app up to the package root
