+++ bxbot-app/pom.xml
@@ -15,7 +15,7 @@
     </parent>
     <properties>
         <springframework.version>4.3.1.RELEASE</springframework.version>
-        <start-class>com.gazbert.bxbot.core.BXBot</start-class>
+        <start-class>com.gazbert.bxbot.BXBot</start-class>
     </properties>
     <dependencies>
         <dependency>
