+++ bxbot-rest-api/src/main/java/com/gazbert/bxbot/rest/api/EngineConfigController.java
@@ -21,9 +21,9 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.gazbert.bxbot.core.rest.endpoints;
+package com.gazbert.bxbot.rest.api;
 
-import com.gazbert.bxbot.core.rest.security.User;
+import com.gazbert.bxbot.rest.security.User;
 import com.gazbert.bxbot.domain.engine.EngineConfig;
 import com.gazbert.bxbot.services.EngineConfigService;
 import org.springframework.beans.factory.annotation.Autowired;
