+++ bxbot-rest-api/src/main/java/com/gazbert/bxbot/rest/security/User.java
@@ -21,25 +21,22 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.gazbert.bxbot.core.rest.security;
+package com.gazbert.bxbot.rest.security;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.hibernate.validator.constraints.NotEmpty;
-import org.springframework.security.core.GrantedAuthority;
 
 import javax.persistence.*;
 import java.util.HashSet;
 import java.util.Set;
 
 /**
- * Encapsulates a role containing selected users of the REST API.
+ * Represents a User of the REST API.
  *
  * @author gazbert
  */
 @Entity
-public class Role implements GrantedAuthority {
-
-    private static final long serialVersionUID = 8224732733933233303L;
+public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
@@ -48,13 +45,28 @@ public class Role implements GrantedAuthority {
     @NotEmpty
     private String name;
 
-    @JsonIgnore
-    @ManyToMany(fetch = FetchType.LAZY, mappedBy = "roles")
-    private Set<User> users = new HashSet<>();
+    @NotEmpty
+    @Column(name = "login_id", unique = true, nullable = false)
+    private String loginId;
 
-    @Override
-    public String getAuthority() {
-        return name;
+    @NotEmpty
+    private String password;
+
+    @JsonIgnore
+    @ManyToMany(fetch = FetchType.EAGER)
+    @JoinTable(name = "user_role", joinColumns = {@JoinColumn(name = "user_id")}, inverseJoinColumns = {@JoinColumn(name = "role_id")})
+    private Set<Role> roles = new HashSet<>();
+
+    public User() {
+    }
+
+    public User(User user) {
+        super();
+        this.id = user.getId();
+        this.name = user.getName();
+        this.loginId = user.getLoginId();
+        this.password = user.getPassword();
+        this.roles = user.getRoles();
     }
 
     public Integer getId() {
@@ -73,11 +85,28 @@ public class Role implements GrantedAuthority {
         this.name = name;
     }
 
-    public Set<User> getUsers() {
-        return users;
+    public String getLoginId() {
+        return loginId;
     }
 
-    public void setUsers(Set<User> users) {
-        this.users = users;
+    public void setLoginId(String loginId) {
+        this.loginId = loginId;
     }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public Set<Role> getRoles() {
+        return roles;
+    }
+
+    public void setRoles(Set<Role> roles) {
+        this.roles = roles;
+    }
+
 }
