+++ bxbot-rest-api/src/main/java/com/gazbert/bxbot/rest/security/WebSecurityConfiguration.java
@@ -21,7 +21,7 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.gazbert.bxbot.core.rest.security;
+package com.gazbert.bxbot.rest.security;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
