+++ bxbot-rest-api/src/test/java/com/gazbert/bxbot/rest/api/AbstractConfigControllerTest.java
@@ -21,9 +21,9 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.gazbert.bxbot.core.rest.endpoints;
+package com.gazbert.bxbot.rest.api;
 
-import com.gazbert.bxbot.core.rest.security.OAuth2ServerConfiguration;
+import com.gazbert.bxbot.rest.security.OAuth2ServerConfiguration;
 import org.junit.Assert;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
